refactor(component) - delete index files
