--- src/components/AboutMe/index.ts
@@ -1,1 +1,0 @@
-export { default as AboutMeSection } from './AboutMeSection/AboutMeSection'
