--- src/components/Projects/index.ts
@@ -1,3 +1,0 @@
-export { default as ProjectItem } from './ProjectItem/ProjectItem'
-export { default as ProjectList } from './ProjectList/ProjectList'
-export { default as ProjectsSection } from './ProjectsSections/ProjectsSection'
